fix bug of multiple chunks
When the chunk was longer than 2Gb it used to be split into two pieces
at random and each piece would be read by fread. The split could happen
in the middle of a line, making fread discard that line and causing
troubles. Now, the chunk is split exactly at a newline.

In addition, we now run gc() before reading a long chunk, so as to
reduce the memory usage.